requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pandas>=2.2.0
scikit-learn>=1.4.0
Pillow>=10.2.0
pyyaml>=6.0.1
matplotlib>=3.8.0
